get latest data, fix no internet, add latest time

+++ app/src/main/java/com/thanhthido/androiddashboard/data/mqtt_response/SensorData.kt
@@ -9,5 +9,5 @@ data class SensorData(
     @SerializedName("type")
     val type: String = "",
     @SerializedName("value")
-    val value: Int = 0
+    val value: Double = 0.0
 )
